refactor(harness-cli): rename create.ts/createCommand -> new.ts/newCommand
Follow-through on the `create` -> `new` verb rename so the code matches
the CLI: rename the command file + symbol and the wizard/test files,
and fix doc-comments that still named the command `create`. A
maintainer searching for the `new` command now finds it.

- src/commands/create.ts -> new.ts (createCommand -> newCommand)
- src/commands/create-wizard.tsx -> new-wizard.tsx (runCreateWizard ->
  runNewWizard)
- test/{create-scaffold,create-wizard}.test.ts -> {new-scaffold,new-wizard}
- doc-comments: prune-targets.ts, model-catalog.ts, command.ts,
  new-wizard.tsx header

No behavior change; 70/70 unit tests, tsc clean, dispatch smoke green.
